Bugfixes: orphan images & double crop
Bugfix: orphan images.
Remove images from disk when deleting gifts, removing or updating images.

Bugfix: double cropping
We were cropping images on the client side, sending the crop down to the server and cropping it again before storing on disk. That led to aweful results as one can imagine. The Sharp library running on the server was basically given junk to work with.