🚸 Support vey narrow screens
We could have gotten away with just updating the CSS using a `@media (max-width: 500px){}`statement but nah. We just had to get the focal points of the image juuuust right... wihch led to an overengineered solution. But hey, we learned a bunch of stuff.

npm overrides

smartcrop-sharp is not frequently maintained and has a peer dependency on sharp. We're overriding it to avoid unnecessary warning.